+++ packages/terra-core-docs/src/terra-dev-site/doc/status-view/example/ToggleStatusView.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import StatusView from 'terra-status-view';
+import ToggleButton from 'terra-toggle-button';
+
+const ToggleStatusView = () => (
+  <ToggleButton closedButtonText="Trigger Error">
+    <StatusView
+      message="run time error"
+      variant="error"
+    />
+  </ToggleButton>
+);
+
+export default ToggleStatusView;