+++ packages/terra-status-view/src/StatusView.jsx
@@ -162,7 +162,7 @@ const StatusView = ({
   ]);
 
   return (
-    <div {...customProps} className={outerViewClassNames}>
+    <div {...customProps} className={outerViewClassNames} role="alert" aria-live="polite">
       <div className={cx('top-space')} />
       <div className={innerViewClassNames}>
         {glyphSection}