+++ packages/terra-core-docs/CHANGELOG.md
@@ -3,11 +3,15 @@
 ## Unreleased
 
 * Added
+
+  * Added dynamic example for terra-status-view.
+
   * Added the UX recommendation for terra-content-container
 
 * Changed
   * Updated icon documentation to incorporate `one-cerner-style-icons` v1.5.0.
   * Updated non-meaningful content with meaningful content for `action-header`, `arrange`, `content-container`, `divider`, `checkbox`, `radio`, `paginator`, `scroll`, `section-header`, `show-hide`, `toggle`, `toggle-button` and `toggle-section-header`
+
 
 * Added
   * Updated Usage Guide to include information on decorative, supportive and meaningful text attributes.