+++ packages/terra-status-view/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added dynamic example for terra-status-view.
+
 ## 4.54.0 - (March 1, 2023)
 
 * Changed